refactor: passa a função adicionarTarefa para o arquivo tarefaLogica e importa para o script.js, reescreve função de receber input da tarefa (bug de retornar função)
