+++ js/script.js
@@ -1,15 +1,24 @@
-import { criaItemTarefa, criaBotaoRemover } from "./modules/manipulaDom.js";
-import { deletarTarefaCallback } from "./modules/tarefaLogica.js";
+import { criaItemTarefa} from "./modules/manipulaDom.js";
+import { adicionarTarefa } from "./modules/tarefaLogica.js";
 
 let contadorIdsTarefas = 0;
 const botaoAdicionarTarefa = document.querySelector(".btn-add") 
 const inputTarefa = document.querySelector("input") 
 const containerListaTarefas = document.querySelector("ul") 
-let arrayTarefas = []
 
-const deletaTarefa = deletarTarefaCallback(containerListaTarefas, arrayTarefas)
 
-botaoAdicionarTarefa.addEventListener("click", adicionarTarefa) 
+
+
+botaoAdicionarTarefa.addEventListener("click", () =>{
+    adicionarTarefa(valorInput, contadorIdsTarefas, containerListaTarefas)
+}) 
+
+
+function valorInput() {
+    const abc = inputTarefa.value.trim() 
+    return  abc
+} 
+
 
 function salvarTarefasLocalstorage(){
     localStorage.setItem("Dados Tarefas", JSON.stringify(arrayTarefas))
@@ -19,9 +28,6 @@ function salvarIdLocalstorage(){
     localStorage.setItem("Contador do ID", JSON.stringify(contadorIdsTarefas))
 }
 
-function recebeValorInput(){
-    return inputTarefa.value.trim() 
-}
 
 function limpaValorInput(){
     inputTarefa.value = ''
@@ -29,21 +35,7 @@ function limpaValorInput(){
 
 
 
-function adicionarTarefa(){
-    const conteudoTarefa = recebeValorInput()
-    if(conteudoTarefa === ''){
-        alert("Coloque uma tarefa válida")
-        return
-    } 
-    contadorIdsTarefas += 1;
-    const itemTarefa = criaItemTarefa(conteudoTarefa, String(contadorIdsTarefas), deletaTarefa)
-    containerListaTarefas.appendChild(itemTarefa)
-    let objetoTarefa = {conteudo: conteudoTarefa, id: String(contadorIdsTarefas)}
-    arrayTarefas.push(objetoTarefa)
-    limpaValorInput()
-    // salvarTarefasLocalstorage()
-    // salvarIdLocalstorage()
-}
+
